requirements.txt
flask==3.0.3
pdfplumber==0.11.4
python-docx==1.1.2
werkzeug==3.0.4
